feat(ai-chat-screen): - finished logic implementation and merged it with the ui from amr branch
