+++ lib/core/routes/routes.dart
@@ -17,8 +17,8 @@ class Routes {
   static const String mealDetails = '/mealDetails';
   static const String exercise = '/exercise';
   static const String mealDetailsScreen = '/mealDetailsScreen';
-  static const String onboardingChatScreen = '/onboardingChatScreen';
-  static const String chatBootScreen = '/chatBootScreen';
+  static const String onboardingSmartCoachScreen = '/onboardingSmartCoachScreen';
+  static const String smartCoachScreen = '/smartCoachScreen';
