+++ lib/feature/chat_ai/data/api/smart_coach_services.dart
@@ -2,8 +2,8 @@ import 'package:flutter_gemini/flutter_gemini.dart';
 import 'package:injectable/injectable.dart';
 @lazySingleton
 class SmartCoachService {
-  Stream<Candidates?> streamChat(List<Content> history) {
-    return Gemini.instance.streamChat(history);
+  Stream<Candidates?> streamChat(List<Content> history,{String? model}) {
+    return Gemini.instance.streamChat(history,modelName: model);
   }
 }
 // class SmartCoachService {
