+++ lib/feature/chat_ai/data/data_source/remote_smart_coach_data_source.dart
@@ -2,5 +2,5 @@
 import 'package:flutter_gemini/flutter_gemini.dart';
 
 abstract class SmartCoachRemoteDataSource {
-  Stream<Candidates?> getSmartCoachResponseStream(List<Content> chatHistory);
+  Stream<Candidates?> getSmartCoachResponseStream(List<Content> chatHistory,{String? model});
 }
